pyproject.toml
[project]
name = "wplace-health-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cloudscraper>=1.2.71",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
